Add TestHelpers.with_puzzle_input/2
It reads the given input filename and executes the given test function
body if the file was found and read. This allows the test suite to be
run even though the private input files aren't present if someone else
tries to run these tests. A remaining flaw is that the tests remain
hardcoded to to my results.